Finish moving elements
