+++ src/elements/emby-itemrefreshindicator/emby-itemrefreshindicator.js
