+++ src/elements/emby-itemscontainer/emby-itemscontainer.js
