+++ src/elements/emby-playstatebutton/emby-playstatebutton.js
