+++ src/elements/emby-programcell/emby-programcell.js
@@ -1,0 +1,16 @@
+define([], function() {
+    'use strict';
+
+    var ProgramCellPrototype = Object.create(HTMLButtonElement.prototype);
+
+    ProgramCellPrototype.detachedCallback = function () {
+        this.posLeft = null;
+        this.posWidth = null;
+        this.guideProgramName = null;
+    };
+
+    document.registerElement('emby-programcell', {
+        prototype: ProgramCellPrototype,
+        extends: 'button'
+    });
+});
