+++ src/elements/emby-ratingbutton/emby-ratingbutton.js
