+++ src/elements/emby-scroller/emby-scroller.js
