+++ src/elements/emby-tabs/emby-tabs.css
