+++ src/elements/emby-tabs/emby-tabs.js
